plugins: change interface
